Coverity fixes 2

+++ picoquic/picoquic_internal.h
@@ -1116,10 +1116,10 @@ void picoquic_update_pacing_rate(picoquic_path_t* path_x, double pacing_rate, ui
 void picoquic_reinsert_by_wake_time(picoquic_quic_t* quic, picoquic_cnx_t* cnx, uint64_t next_time);
 
 /* Integer parsing macros */
-#define PICOPARSE_16(b) ((((uint16_t)(b)[0]) << 8) | (b)[1])
-#define PICOPARSE_24(b) ((((uint32_t)PICOPARSE_16(b)) << 8) | ((b)[2]))
-#define PICOPARSE_32(b) ((((uint32_t)PICOPARSE_16(b)) << 16) | PICOPARSE_16((b) + 2))
-#define PICOPARSE_64(b) ((((uint64_t)PICOPARSE_32(b)) << 32) | PICOPARSE_32((b) + 4))
+#define PICOPARSE_16(b) ((((uint16_t)(b)[0]) << 8) | (uint16_t)((b)[1]))
+#define PICOPARSE_24(b) ((((uint32_t)PICOPARSE_16(b)) << 8) | (uint32_t)((b)[2]))
+#define PICOPARSE_32(b) ((((uint32_t)PICOPARSE_16(b)) << 16) | (uint32_t)PICOPARSE_16((b) + 2))
+#define PICOPARSE_64(b) ((((uint64_t)PICOPARSE_32(b)) << 32) | (uint64_t)PICOPARSE_32((b) + 4))
 
 /* Integer formatting functions */
 void picoformat_16(uint8_t* bytes, uint16_t n16);
